chore: move leaflet handling deps
