+++ src/blocks/frontend/leaflet-map/language-content.js
